# docs/swagger.yaml
definitions:
  errorcode.Error:
    type: object
...
  requests.IngressPathRule:
    properties:
      path:
        description: 访问路径（如 /、/api、/api/(.*)）
        type: string
      path_type:
        description: 'pathType: Exact / Prefix / ImplementationSpecific'
        type: string
      service_name:
        description: 后端 Service 名称
        type: string
      service_port:
        description: 后端 Service 端口（支持命名端口，故用字符串；可传 "80" 或 "http"）
        type: string
    type: object
  requests.IngressRule:
    properties:
      host:
        type: string
      paths:
        items:
          $ref: '#/definitions/requests.IngressPathRule'
        type: array
    type: object
  requests.IngressTLS:
    properties:
      hosts:
        items:
          type: string
        type: array
      secret_name:
        type: string
    type: object
...
  requests.KubeIngressCreateRequest:
    properties:
      annotations:
        description: 可选注解（如 rewrite、限流等）
        type: string
      ingress_class_name:
        description: 例如："nginx"、"traefik"；可为空（走默认 class）
        type: string
      labels:
        description: 可选标签
        type: string
      name:
        type: string
      namespace:
        type: string
      rules:
        description: 规则：支持多个 host
        items:
          $ref: '#/definitions/requests.IngressRule'
        type: array
      tls:
        description: 可选 TLS 列表
        items:
          $ref: '#/definitions/requests.IngressTLS'
        type: array
    type: object
...
info:
  contact: {}
  description: 基于Gin+Vue开发的K8s管理平台
  termsOfService: https://gitee.com/jay-kim/k8s_operation
  title: K8s管理平台
  version: "1.0"
paths:
...
  /api/v1/k8s/ingress/apply-yaml:
    put:
      consumes:
      - application/json
      parameters:
      - description: YAML 内容
        in: body
        name: body
        required: true
        schema:
          properties:
            yaml:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 应用 Ingress YAML
      tags:
      - K8s Ingress 管理
  /api/v1/k8s/ingress/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.KubeIngressCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 创建 Ingress
      tags:
      - K8s Ingress 管理
  /api/v1/k8s/ingress/create-from-yaml:
    post:
      consumes:
      - application/json
      parameters:
      - description: YAML 内容
        in: body
        name: body
        required: true
        schema:
          properties:
            yaml:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 从 YAML 创建 Ingress
      tags:
      - K8s Ingress 管理
  /api/v1/k8s/ingress/delete:
    delete:
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: Ingress 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 删除 Ingress
      tags:
      - K8s Ingress 管理
  /api/v1/k8s/ingress/detail:
    get:
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: Ingress 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 获取 Ingress 详情
      tags:
      - K8s Ingress 管理
  /api/v1/k8s/ingress/list:
    get:
      description: 支持分页、命名空间过滤、名称模糊查询
      parameters:
      - description: 命名空间
        in: query
        maxLength: 100
        name: namespace
        type: string
      - description: Ingress 名称(模糊匹配)
        in: query
        maxLength: 100
        name: name
        type: string
      - description: 页码 (从1开始)
        in: query
        name: page
        required: true
        type: integer
      - description: 每页数量 (默认20)
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 获取 K8s Ingress 列表
      tags:
      - K8s Ingress 管理
  /api/v1/k8s/ingress/patch:
    patch:
      consumes:
      - application/strategic-merge-patch+json
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: Ingress 名称
        in: query
        name: name
        required: true
        type: string
      - description: Patch Body(JSON字符串)
        in: body
        name: content
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: Patch Ingress（StrategicMergePatch）
      tags:
      - K8s Ingress 管理
  /api/v1/k8s/ingress/patch-json:
    post:
      consumes:
      - application/merge-patch+json
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: Ingress 名称
        in: query
        name: name
        required: true
        type: string
      - description: Patch Body(JSON字符串)
        in: body
        name: content
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: Patch Ingress（JSON Merge Patch – 覆盖式）
      tags:
      - K8s Ingress 管理
  /api/v1/k8s/ingress/yaml:
    get:
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: Ingress 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: YAML 内容
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 获取 Ingress 的 YAML
      tags:
      - K8s Ingress 管理
...
swagger: "2.0"